Moved landingpage react component into resources folder so that I would be able to run next.js production build function
