+++ src/pages/assets/button.js
@@ -3,6 +3,8 @@ import Link from 'next/link';
 
 
 export default function Button(props) {
+  props.link = props.link || '';
+  props.name = props.name || '';
   return (
     <button className={`hover:${props.hoverColor} ${props.color} text-white font-bold py-2 px-4 rounded mb-5 tracking-wide min-w-fit max-w-fit md:max-w-full mx-auto md:mx-0`}>
       <Link href={props.link}>{props.name}</Link>
